[ADD/#5] 뷰모델과 상태 파일 분리
